+++ django_deovi/admin/directory.py
@@ -10,4 +10,12 @@ from ..models import Directory
 
 @admin.register(Directory)
 class DirectoryAdmin(SmartModelAdmin):
-    pass
+    list_display = (
+        "path",
+        "title",
+        "device",
+        "created_date",
+        "last_update",
+    )
+    list_filter = ("device", "created_date", "last_update")
+    search_fields = ["path", "title"]