+++ django_deovi/admin/media.py
@@ -6,9 +6,14 @@ from django.contrib import admin
 from ..models import MediaFile
 
 
+@admin.register(MediaFile)
 class MediaFileAdmin(admin.ModelAdmin):
-    pass
-
-
-# Registering interface to model
-admin.site.register(MediaFile, MediaFileAdmin)
+    list_display = (
+        "filename",
+        "directory",
+        "container",
+        "stored_date",
+        "loaded_date",
+    )
+    list_filter = ("container", "stored_date", "loaded_date")
+    search_fields = ["filename", "path"]